more update

+++ pkg/facet/number_field_test.go
@@ -12,8 +12,8 @@ func makeNumberField[K float64 | int]() *NumberField[K] {
 	r := EmptyNumberField[K](&BaseField{Id: 1, Name: "number", Description: "number field"})
 	for i := 0; i < total; i++ {
 		for j := 0; j < 100; j++ {
-			r.AddValueLink(K(i), (total*100)+j)
-			r.AddValueLink(K(i), (total*100)+total+i+j)
+			r.AddValueLink(K(i), uint((total*100)+j))
+			r.AddValueLink(K(i), uint((total*100)+total+i+j))
 		}
 	}
 	return r
